Added More Test Cases

+++ src/main/java/com/example/demo/TwoSum.java
@@ -14,13 +14,13 @@ public class TwoSum {
         // // Your solution will go here
         // return new int[0]; // Placeholder return
 
-        Map<Integer, Integer> map = new HashMap<>();
+        Map<Long, Integer> map = new HashMap<>();
         for (int i = 0; i < nums.length; i++) {
-            int complement = target - nums[i];
+            long complement = (long) target - nums[i];
             if (map.containsKey(complement)) {
                 return new int[] { map.get(complement), i };
             }
-            map.put(nums[i], i);
+            map.put((long)nums[i], i); // Corrected line: cast nums[i] to long
         }
         throw new IllegalArgumentException("No two sum solution");
     }
